backend/pyproject.toml
[project]
name = "rag-website-embedding"
version = "0.1.0"
description = "RAG system for embedding documentation website content"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "openai>=1.0.0",
    "qdrant-client>=1.7.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.5.0",
    "pytest>=7.4.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"